@@ -36,15 +36,15 @@
     "danger": "^10.6.4",
     "date-fns": "^1.30.1",
     "dictionary-en": "^3.1.0",
-    "eslint": "^7.21.0",
+    "eslint": "^7.24.0",
     "eslint-config-google": "^0.14.0",
-    "eslint-config-prettier": "^8.1.0",
+    "eslint-config-prettier": "^8.2.0",
     "eslint-plugin-filenames": "^1.3.2",
-    "eslint-plugin-flowtype": "^5.3.1",
+    "eslint-plugin-flowtype": "^5.7.1",
     "eslint-plugin-import": "^2.22.1",
     "eslint-plugin-jsx-a11y": "^6.4.1",
-    "eslint-plugin-prettier": "^3.3.1",
-    "eslint-plugin-react": "^7.22.0",
+    "eslint-plugin-prettier": "^3.4.0",
+    "eslint-plugin-react": "^7.23.2",
     "fs-extra": "^8.1.0",
     "glob": "^7.1.6",
     "husky": "3.1.0",
