@@ -1,6 +1,6 @@
 {
   "name": "create-gatsby",
-  "version": "1.4.0-next.0",
+  "version": "1.5.0-next.0",
   "main": "lib/index.js",
   "bin": "cli.js",
   "license": "MIT",
@@ -24,10 +24,10 @@
     "ansi-wordwrap": "^1.0.2",
     "common-tags": "^1.8.0",
     "enquirer": "^2.3.6",
-    "eslint": "^7.18.0",
+    "eslint": "^7.24.0",
     "execa": "^4.1.0",
     "fs-extra": "^9.0.1",
-    "gatsby-plugin-utils": "^1.4.0-next.0",
+    "gatsby-plugin-utils": "^1.5.0-next.0",
     "joi": "^17.2.1",
     "microbundle": "^0.13.0",
     "node-fetch": "^2.6.1",
