@@ -1,7 +1,7 @@
 {
   "name": "gatsby",
   "description": "Blazing fast modern site generator for React",
-  "version": "3.4.0-next.3",
+  "version": "3.5.0-next.0",
   "author": "Kyle Mathews <mathews.kyle@gmail.com>",
   "bin": {
     "gatsby": "./cli.js"
@@ -34,8 +34,8 @@
     "babel-plugin-add-module-exports": "^1.0.4",
     "babel-plugin-dynamic-import-node": "^2.3.3",
     "babel-plugin-lodash": "^3.3.4",
-    "babel-plugin-remove-graphql-queries": "^3.4.0-next.1",
-    "babel-preset-gatsby": "^1.4.0-next.1",
+    "babel-plugin-remove-graphql-queries": "^3.5.0-next.0",
+    "babel-preset-gatsby": "^1.5.0-next.0",
     "better-opn": "^2.0.0",
     "better-queue": "^3.8.10",
     "bluebird": "^3.7.2",
@@ -46,6 +46,7 @@
     "chokidar": "^3.4.2",
     "common-tags": "^1.8.0",
     "compression": "^1.7.4",
+    "cookie": "^0.4.1",
     "copyfiles": "^2.3.0",
     "core-js": "^3.9.0",
     "cors": "^2.8.5",
@@ -77,16 +78,16 @@
     "find-cache-dir": "^3.3.1",
     "fs-exists-cached": "1.0.0",
     "fs-extra": "^8.1.0",
-    "gatsby-cli": "^3.4.0-next.2",
-    "gatsby-core-utils": "^2.4.0-next.1",
-    "gatsby-graphiql-explorer": "^1.4.0-next.0",
-    "gatsby-legacy-polyfills": "^1.4.0-next.0",
-    "gatsby-link": "^3.4.0-next.0",
-    "gatsby-plugin-page-creator": "^3.4.0-next.1",
-    "gatsby-plugin-typescript": "^3.4.0-next.1",
-    "gatsby-plugin-utils": "^1.4.0-next.0",
-    "gatsby-react-router-scroll": "^4.4.0-next.0",
-    "gatsby-telemetry": "^2.4.0-next.1",
+    "gatsby-cli": "^3.5.0-next.0",
+    "gatsby-core-utils": "^2.5.0-next.0",
+    "gatsby-graphiql-explorer": "^1.5.0-next.0",
+    "gatsby-legacy-polyfills": "^1.5.0-next.0",
+    "gatsby-link": "^3.5.0-next.0",
+    "gatsby-plugin-page-creator": "^3.5.0-next.0",
+    "gatsby-plugin-typescript": "^3.5.0-next.0",
+    "gatsby-plugin-utils": "^1.5.0-next.0",
+    "gatsby-react-router-scroll": "^4.5.0-next.0",
+    "gatsby-telemetry": "^2.5.0-next.0",
     "glob": "^7.1.6",
     "got": "8.3.2",
     "graphql": "^15.4.0",
@@ -119,6 +120,7 @@
     "opentracing": "^0.14.4",
     "p-defer": "^3.0.0",
     "parseurl": "^1.3.3",
+    "path-to-regexp": "0.1.7",
     "physical-cpu-count": "^2.0.0",
     "platform": "^1.3.6",
     "pnp-webpack-plugin": "^1.6.4",
@@ -154,7 +156,7 @@
     "util.promisify": "^1.0.1",
     "uuid": "3.4.0",
     "v8-compile-cache": "^2.2.0",
-    "webpack": "^5.28.0",
+    "webpack": "^5.35.0",
     "webpack-dev-middleware": "^4.1.0",
     "webpack-dev-server": "^3.11.2",
     "webpack-merge": "^5.7.3",
@@ -175,7 +177,7 @@
     "@types/string-similarity": "^3.0.0",
     "@types/tmp": "^0.2.0",
     "@types/webpack-virtual-modules": "^0.1.1",
-    "babel-preset-gatsby-package": "^1.4.0-next.0",
+    "babel-preset-gatsby-package": "^1.5.0-next.0",
     "cross-env": "^7.0.3",
     "documentation": "^13.1.0",
     "enhanced-resolve": "^4.2.0",
