@@ -11,9 +11,9 @@
   "license": "0BSD",
   "dependencies": {
     "@mdx-js/react": "^1.6.22",
-    "gatsby": "^2.32.12",
-    "gatsby-theme-blog-core": "^2.0.4",
-    "react": "^16.12.0",
-    "react-dom": "^16.12.0"
+    "gatsby": "^3.3.1",
+    "gatsby-theme-blog-core": "^3.0.0",
+    "react": "^17.0.2",
+    "react-dom": "^17.0.2"
   }
 }
