@@ -10,10 +10,10 @@
   },
   "license": "0BSD",
   "dependencies": {
-    "gatsby": "^2.32.12",
-    "gatsby-theme-blog": "^2.0.2",
-    "react": "^16.12.0",
-    "react-dom": "^16.12.0",
-    "theme-ui": "0.4.0-rc.14"
+    "gatsby": "^3.3.1",
+    "gatsby-theme-blog": "^3.0.0",
+    "react": "^17.0.2",
+    "react-dom": "^17.0.2",
+    "theme-ui": "0.7.0"
   }
 }
