@@ -9,9 +9,9 @@
     "build": "gatsby build"
   },
   "dependencies": {
-    "gatsby": "^2.32.12",
+    "gatsby": "^3.3.0",
     "gatsby-theme-minimal": "^1.0.0",
-    "react": "^16.12.0",
-    "react-dom": "^16.12.0"
+    "react": "^17.0.2",
+    "react-dom": "^17.0.2"
   }
 }
