@@ -1,31 +1,19 @@
-import React from "react"
-
+import * as React from "react"
 import emitter from "../emitter"
 import { Indicator } from "./indicator"
 
-// no hooks because we support react versions without hooks support
-export class LoadingIndicatorEventHandler extends React.Component {
-  state = { visible: false }
+export function LoadingIndicatorEventHandler() {
+  const [visible, setVisible] = React.useState(false)
 
-  show = () => {
-    this.setState({ visible: true })
-  }
+  React.useEffect(() => {
+    emitter.on(`onDelayedLoadPageResources`, () => setVisible(true))
+    emitter.on(`onRouteUpdate`, () => setVisible(false))
 
-  hide = () => {
-    this.setState({ visible: false })
-  }
+    return () => {
+      emitter.off(`onDelayedLoadPageResources`, () => setVisible(true))
+      emitter.off(`onRouteUpdate`, () => setVisible(false))
+    }
+  }, [])
 
-  componentDidMount() {
-    emitter.on(`onDelayedLoadPageResources`, this.show)
-    emitter.on(`onRouteUpdate`, this.hide)
-  }
-
-  componentWillUnmount() {
-    emitter.off(`onDelayedLoadPageResources`, this.show)
-    emitter.off(`onRouteUpdate`, this.hide)
-  }
-
-  render() {
-    return <Indicator visible={this.state.visible} />
-  }
+  return <Indicator visible={visible} />
 }
