@@ -1,5 +1,7 @@
-export const onCreateBabelConfig = ({ actions }, pluginOptions) => {
+export const onCreateBabelConfig = ({ actions }) => {
   actions.setBabelPreset({
     name: require.resolve(`@babel/preset-flow`),
   })
 }
+
+export const pluginOptionsSchema = ({ Joi }) => Joi.object({})
