@@ -4,7 +4,7 @@ import { createPath } from "gatsby-page-utils"
 import { Reporter } from "gatsby/reporter"
 import { Options as ISlugifyOptions } from "@sindresorhus/slugify"
 import { reverseLookupParams } from "./extract-query"
-import { getMatchPath } from "./get-match-path"
+import { getMatchPath } from "gatsby-core-utils"
 import { getCollectionRouteParams } from "./get-collection-route-params"
 import { derivePath } from "./derive-path"
 import { watchCollectionBuilder } from "./watch-collection-builder"
@@ -105,9 +105,12 @@ ${errors.map(error => error.message).join(`\n`)}`.trim(),
 
   let derivePathErrors = 0
 
+  const knownPagePaths = new Set<string>()
+
   // 3. Loop through each node and create the page, also save the path it creates to pass to the watcher
   //    the watcher will use this data to delete the pages if the query changes significantly.
-  const paths = nodes.map((node: Record<string, Record<string, unknown>>) => {
+  const paths: Array<string> = []
+  nodes.forEach((node: Record<string, Record<string, unknown>>) => {
     // URL path for the component and node
     const { derivedPath, errors } = derivePath(
       filePath,
@@ -116,6 +119,11 @@ ${errors.map(error => error.message).join(`\n`)}`.trim(),
       slugifyOptions
     )
     const path = createPath(derivedPath)
+    // We've already created a page with this path
+    if (knownPagePaths.has(path)) {
+      return
+    }
+    knownPagePaths.add(path)
     // Params is supplied to the FE component on props.params
     const params = getCollectionRouteParams(createPath(filePath), path)
     // nodeParams is fed to the graphql query for the component
@@ -135,7 +143,7 @@ ${errors.map(error => error.message).join(`\n`)}`.trim(),
 
     derivePathErrors += errors
 
-    return path
+    paths.push(path)
   })
 
   if (derivePathErrors > 0) {
