@@ -9,14 +9,12 @@ exports.setFieldsOnGraphQLNodeType = require(`./extend-node-type`)
 
 exports.pluginOptionsSchema = function ({ Joi }) {
   return Joi.object({
-    commonmark: Joi.boolean().description(
-      `Activates CommonMark mode (default: true)`
-    ),
+    // Options `commonmark` and `pedantic` have no effect since gatsby-transformer-remark@4.0.0
+    // TODO: remove in v5
+    commonmark: Joi.any(),
+    pedantic: Joi.any(),
     footnotes: Joi.boolean().description(
       `Activates Footnotes mode (default: true)`
-    ),
-    pedantic: Joi.boolean().description(
-      `Activates pedantic mode (default: true)`
     ),
     gfm: Joi.boolean().description(
       `Activates GitHub Flavored Markdown mode (default: true)`
