@@ -1,12 +1,5 @@
 import * as React from "react"
-
-function css(strings, ...keys) {
-  const lastIndex = strings.length - 1
-  return (
-    strings.slice(0, lastIndex).reduce((p, s, i) => p + s + keys[i], ``) +
-    strings[lastIndex]
-  )
-}
+import { css } from "../css-to-object"
 
 export const Style = () => (
   <style
@@ -46,8 +39,8 @@ export const Style = () => (
           --codeFrame-color: #414141;
           --codeFrame-button-bg: white;
           --radii: 5px;
-          --z-index-backdrop: 8000;
-          --z-index-overlay: 9000;
+          --z-index-backdrop: 9000;
+          --z-index-overlay: 10000;
           --space: 1.5em;
           --space-sm: 1em;
           --space-lg: 2.5em;
