@@ -1,28 +1,17 @@
 import React from "react"
+import { css } from "../css-to-object"
 
-function css(strings, ...keys) {
-  const lastIndex = strings.length - 1
-  return (
-    strings.slice(0, lastIndex).reduce((p, s, i) => p + s + keys[i], ``) +
-    strings[lastIndex]
-  )
-}
-
-const Style = () => (
+export const Style = () => (
   <style
     dangerouslySetInnerHTML={{
       __html: css`
         :host {
-          --purple-60: #663399;
-          --gatsby: var(--purple-60);
-          --purple-40: #b17acc;
-          --purple-20: #f1defa;
-          --dimmedWhite: rgba(255, 255, 255, 0.8);
-          --white: #ffffff;
-          --black: #000000;
-          --grey-90: #232129;
+          --spinnerColor: #663399;
+          --borderLeft: #b17acc;
+          --background: #ffffff;
+          --color: #232129;
           --radii: 4px;
-          --z-index-indicator: 9000;
+          --z-index-indicator: 10000;
           --shadow: 0px 2px 4px rgba(46, 41, 51, 0.08),
             0px 4px 8px rgba(71, 63, 79, 0.16);
         }
@@ -31,15 +20,15 @@ const Style = () => (
           font: 14px/1.5 -apple-system, BlinkMacSystemFont, "Segoe UI", Roboto,
             Helvetica, Arial, sans-serif, "Apple Color Emoji", "Segoe UI Emoji",
             "Segoe UI Symbol" !important;
-          background: var(--white);
-          color: var(--grey-90);
+          background: var(--background);
+          color: var(--color);
           position: fixed;
           bottom: 1.5em;
           left: 1.5em;
           box-shadow: var(--shadow);
           border-radius: var(--radii);
           z-index: var(--z-index-indicator);
-          border-left: 0.25em solid var(--purple-40);
+          border-left: 0.25em solid var(--borderLeft);
           display: flex;
           align-items: center;
           justify-content: space-between;
@@ -67,7 +56,7 @@ const Style = () => (
           animation: spin 1s linear infinite;
           height: 18px;
           width: 18px;
-          color: var(--gatsby);
+          color: var(--spinnerColor);
         }
 
         [data-gatsby-loading-indicator="text"] {
@@ -98,17 +87,13 @@ const Style = () => (
         }
 
         @media (prefers-color-scheme: dark) {
-          [data-gatsby-loading-indicator="root"] {
-            background: var(--grey-90);
-            color: var(--white);
-          }
-          [data-gatsby-loading-indicator="spinner"] {
-            color: var(--purple-20);
+          :host {
+            --spinnerColor: #f1defa;
+            --background: #232129;
+            --color: #ffffff;
           }
         }
       `,
     }}
   />
 )
-
-export default Style
