@@ -8,6 +8,8 @@ import { match as reachMatch } from "@gatsbyjs/reach-router/lib/utils"
 import onExit from "signal-exit"
 import report from "gatsby-cli/lib/reporter"
 import multer from "multer"
+import pathToRegexp from "path-to-regexp"
+import cookie from "cookie"
 
 import telemetry from "gatsby-telemetry"
 
@@ -112,32 +114,78 @@ module.exports = async (program: IServeProgram): Promise<void> => {
   const matchPaths = await readMatchPaths(program)
   router.use(matchPathRouter(matchPaths, { root }))
 
-  if (process.env.GATSBY_EXPERIMENTAL_FUNCTIONS) {
-    const compiledFunctionsDir = path.join(
-      program.directory,
-      `.cache`,
-      `functions`
-    )
+  const compiledFunctionsDir = path.join(
+    program.directory,
+    `.cache`,
+    `functions`
+  )
 
-    const functionsManifest = JSON.parse(
+  let functions
+  try {
+    functions = JSON.parse(
       fs.readFileSync(path.join(compiledFunctionsDir, `manifest.json`))
     )
+  } catch (e) {
+    // ignore
+  }
 
+  if (functions) {
     app.use(
       `/api/*`,
       multer().none(),
       express.urlencoded({ extended: true }),
+      (req, res, next) => {
+        const cookies = req.headers.cookie
+
+        if (!cookies) {
+          return next()
+        }
+
+        req.cookies = cookie.parse(cookies)
+
+        return next()
+      },
       express.text(),
       express.json(),
       express.raw(),
       async (req, res, next) => {
-        const { "0": functionName } = req.params
+        const { "0": pathFragment } = req.params
 
-        if (functionsManifest[functionName]) {
+        // Check first for exact matches.
+        let functionObj = functions.find(
+          ({ apiRoute }) => apiRoute === pathFragment
+        )
+
+        if (!functionObj) {
+          // Check if there's any matchPaths that match.
+          // We loop until we find the first match.
+          functions.some(f => {
+            let exp
+            const keys = []
+            if (f.matchPath) {
+              exp = pathToRegexp(f.matchPath, keys)
+            }
+            if (exp && exp.exec(pathFragment) !== null) {
+              functionObj = f
+              const matches = [...pathFragment.match(exp)].slice(1)
+              const newParams = {}
+              matches.forEach(
+                (match, index) => (newParams[keys[index].name] = match)
+              )
+              req.params = newParams
+
+              return true
+            } else {
+              return false
+            }
+          })
+        }
+
+        if (functionObj) {
+          const pathToFunction = functionObj.absoluteCompiledFilePath
           const start = Date.now()
 
           try {
-            const pathToFunction = functionsManifest[functionName]
             delete require.cache[require.resolve(pathToFunction)]
             const fn = require(pathToFunction)
 
@@ -151,7 +199,9 @@ module.exports = async (program: IServeProgram): Promise<void> => {
 
           const end = Date.now()
           console.log(
-            `Executed function "/api/${functionName}" in ${end - start}ms`
+            `Executed function "/api/${functionObj.apiRoute}" in ${
+              end - start
+            }ms`
           )
 
           return
