@@ -57,14 +57,43 @@ function followVariableDeclarations(binding) {
   return binding
 }
 
+function referencesGatsby(path, callee, calleeName) {
+  // This works for es6 imports
+  if (callee.referencesImport(`gatsby`, ``)) {
+    return true
+  } else {
+    // This finds where userStaticQuery was declared and then checks
+    // if it is a "require" and "gatsby" is the argument.
+    const declaration = path.scope.getBinding(calleeName)
+    if (
+      declaration &&
+      declaration.path.node.init?.callee.name === `require` &&
+      declaration.path.node.init.arguments[0].value === `gatsby`
+    ) {
+      return true
+    } else {
+      return false
+    }
+  }
+}
+
 function isUseStaticQuery(path) {
-  return (
-    (path.node.callee.type === `MemberExpression` &&
-      path.node.callee.property.name === `useStaticQuery` &&
-      path.get(`callee`).get(`object`).referencesImport(`gatsby`)) ||
-    (path.node.callee.name === `useStaticQuery` &&
-      path.get(`callee`).referencesImport(`gatsby`))
-  )
+  const callee = path.node.callee
+  if (callee.type === `MemberExpression`) {
+    const property = callee.property
+    if (property.name === `useStaticQuery`) {
+      return referencesGatsby(
+        path,
+        path.get(`callee`).get(`object`),
+        path.node?.callee.object.name
+      )
+    }
+    return false
+  }
+  if (callee.name === `useStaticQuery`) {
+    return referencesGatsby(path, path.get(`callee`), path.node?.callee.name)
+  }
+  return false
 }
 
 const warnForUnknownQueryVariable = (varName, file, usageFunction) =>
