@@ -76,6 +76,7 @@ const developConfig: MachineConfig<IBuildContext, any, AnyEventObject> = {
             parentSpan,
             store,
             webhookBody,
+            shouldRunCreatePagesStatefully: true,
             deferNodeMutation: true,
           }
         },
@@ -292,6 +293,7 @@ const developConfig: MachineConfig<IBuildContext, any, AnyEventObject> = {
             webhookSourcePluginName,
             refresh: true,
             deferNodeMutation: true,
+            shouldRunCreatePagesStatefully: false,
           }
         },
         onDone: {
@@ -320,7 +322,12 @@ const developConfig: MachineConfig<IBuildContext, any, AnyEventObject> = {
         id: `recreate-pages`,
         src: `recreatePages`,
         data: ({ parentSpan, store }: IBuildContext): IDataLayerContext => {
-          return { parentSpan, store, deferNodeMutation: true }
+          return {
+            parentSpan,
+            store,
+            deferNodeMutation: true,
+            shouldRunCreatePagesStatefully: false,
+          }
         },
         onDone: {
           actions: `assignServiceResult`,
