@@ -3,7 +3,7 @@ import { cpuCoreCount } from "gatsby-core-utils"
 
 export const create = (): Worker =>
   new Worker(require.resolve(`./child`), {
-    numWorkers: cpuCoreCount(),
+    numWorkers: Math.max(1, cpuCoreCount() - 1),
     forkOptions: {
       silent: false,
     },
