@@ -1,10 +1,4 @@
 module.exports = {
-  plugins: [
-    {
-      resolve: `gatsby-theme-blog`,
-      options: {},
-    },
-  ],
   // Customize your site metadata:
   siteMetadata: {
     title: `My Blog Title`,
@@ -21,4 +15,10 @@ module.exports = {
       },
     ],
   },
+  plugins: [
+    {
+      resolve: `gatsby-theme-blog`,
+      options: {},
+    },
+  ],
 }
