@@ -5,13 +5,15 @@
  */
 
 const darkBlue = `#007acc`
-const lightBlue = `#66E0FF`
+// const lightBlue = `#66E0FF`
 const blueGray = `#282c35`
 
-export default {
+const theme = {
     colors: {
         text: blueGray,
         primary: darkBlue,
         heading: blueGray,
     },
 }
+
+export default theme
