@@ -1,5 +1,5 @@
 import React from "react"
-import { Styled } from "theme-ui"
+import { Themed } from "theme-ui"
 
 /**
  * Change the content to add your own bio
@@ -8,7 +8,7 @@ import { Styled } from "theme-ui"
 export default function Bio() {
   return (
     <>
-      This is where <Styled.a href="http://example.com/">your name</Styled.a>
+      This is where <Themed.a href="http://example.com/">your name</Themed.a>
       {` `}
       goes.
       <br />
