@@ -1,11 +1,11 @@
 module.exports = {
+  siteMetadata: {
+    title: `Site title`,
+  },
   plugins: [
     {
       resolve: `gatsby-theme-notes`,
       options: {},
     },
   ],
-  siteMetadata: {
-    title: `Site title`,
-  },
 }
