@@ -1,4 +1,6 @@
-import React from "react"
+import * as React from "react"
 import BlogLayout from "gatsby-theme-blog/src/components/layout"
 
-export default props => <BlogLayout {...props}>{props.children}</BlogLayout>
+const Layout = (props) => <BlogLayout {...props}>{props.children}</BlogLayout>
+
+export default Layout
